changing float to double
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@51017 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ retex/editor-base/src/main/java/com/himamis/retex/editor/share/controller/MathFieldController.java
@@ -18,7 +18,7 @@ public class MathFieldController {
     private MathField mathField;
     private TeXSerializer texSerializer;
 
-    private float size = 16;
+    private double size = 16;
     private int type = TeXFormula.SERIF;
 
 	private GraphicsStub graphics;
@@ -29,7 +29,7 @@ public class MathFieldController {
 		graphics = new GraphicsStub();
 	}
 
-    public void setSize(float size) {
+    public void setSize(double size) {
         this.size = size;
     }
